clean: format & cleanup code
